+++ @xen-orchestra/immutable-backups/src/_fileIndex.integ.mts
@@ -2,7 +2,7 @@ import { describe, it } from 'node:test'
 import assert from 'node:assert/strict'
 import fs from 'node:fs/promises'
 import path from 'node:path'
-import * as FileIndex from './fileIndex.mjs'
+import * as FileIndex from './_fileIndex.mjs'
 import * as Directory from './directory.mjs'
 import { tmpdir } from 'node:os'
 import { rimraf } from 'rimraf'
