+++ @xen-orchestra/immutable-backups/src/file.mts
@@ -1,5 +1,5 @@
 import execa from 'execa'
-import { unindexFile, indexFile } from './fileIndex.mjs'
+import { unindexFile, indexFile } from './_fileIndex.mjs'
 import { createLogger } from '@xen-orchestra/log'
 
 // this work only on linux like systems
